添加管理员管理

+++ src/IEManageSystem.Core/Entitys/Authorization/Users/UserManager.cs
@@ -1,5 +1,6 @@
 ﻿using Abp.Domain.Repositories;
 using Abp.Domain.Services;
+using Abp.Domain.Uow;
 using IEManageSystem.Entitys.Authorization.Roles;
 using IEManageSystem.Help.Exceptions;
 using System;
@@ -15,16 +16,33 @@ namespace IEManageSystem.Entitys.Authorization.Users
     {
         public IRepository<User> UserRepository { get; private set; }
 
+        private IUnitOfWorkManager _unitOfWorkManager { get; set; }
+
         private IRepository<Role> _roleRepository { get; set; }
+
+        public bool AutoSaveChanges { get; set; } = true;
 
         public UserManager(
             IRepository<User> userRepository,
-            IRepository<Role> roleRepository
+            IRepository<Role> roleRepository,
+            IUnitOfWorkManager unitOfWorkManager
             )
         {
             UserRepository = userRepository;
 
             _roleRepository = roleRepository;
+
+            _unitOfWorkManager = unitOfWorkManager;
+        }
+
+        protected Task SaveChanges()
+        {
+            if (!AutoSaveChanges || _unitOfWorkManager.Current == null)
+            {
+                return Task.CompletedTask;
+            }
+
+            return _unitOfWorkManager.Current.SaveChangesAsync();
         }
 
         /// <summary>
@@ -37,6 +55,29 @@ namespace IEManageSystem.Entitys.Authorization.Users
             return UserRepository.Get(id);
         }
 
+        public User GetUserForUserName(string userName)
+        {
+            return UserRepository.FirstOrDefault(e=>e.UserName == userName);
+        }
+
+        /// <summary>
+        /// 验证密码
+        /// </summary>
+        /// <param name="user"></param>
+        /// <param name="password"></param>
+        /// <returns></returns>
+        public bool ValidatePassword(User user, string password)
+        {
+            // 验证密码
+            password = Encrypt.MD5Utf8(password);
+
+            if (user.Password == password) {
+                return true;
+            }
+
+            return false;
+        }
+
         /// <summary>
         /// 添加用户
         /// </summary>
@@ -45,7 +86,7 @@ namespace IEManageSystem.Entitys.Authorization.Users
         /// <param name="name"></param>
         /// <param name="tenantId"></param>
         /// <returns></returns>
-        public User AddUser(string userName, string password, string name = null, int? tenantId = null)
+        public async Task<User> CreateUser(string userName, string password, string name = null, int? tenantId = null)
         {
             if ((UserRepository.GetAllList(e => e.UserName == userName)).Any())
             {
@@ -63,8 +104,20 @@ namespace IEManageSystem.Entitys.Authorization.Users
             };
 
             UserRepository.Insert(user);
+            await SaveChanges();
 
             return user;
+        }
+
+        /// <summary>
+        /// 更新密码
+        /// </summary>
+        /// <param name="password"></param>
+        public void UpdatePassword(User user, string password)
+        {
+            password = Encrypt.MD5Utf8(password);
+
+            user.Password = password;
         }
 
         /// <summary>
@@ -72,6 +125,15 @@ namespace IEManageSystem.Entitys.Authorization.Users
         /// </summary>
         public void DeleteUser(int id)
         {
+            var user = UserRepository.FirstOrDefault(id);
+
+            if (user == null)
+            {
+                throw new MessageException("找不到要删除的用户");
+            }
+
+            UserRepository.EnsureCollectionLoaded(user, e => e.UserRoles);
+
             UserRepository.Delete(id);
         }
